+++ test/support/live_views/params.ex
@@ -4,8 +4,6 @@ defmodule Phoenix.LiveViewTest.ParamCounterLive do
   def render(assigns) do
     ~H"""
     The value is: <%= @val %>
-    connect: <%= inspect(@connect_params) %>
-    info: <%= inspect(@connect_info) %>
     mount: <%= inspect(@mount_params) %>
     params: <%= inspect(@params) %>
     """
@@ -19,8 +17,6 @@ defmodule Phoenix.LiveViewTest.ParamCounterLive do
        socket,
        val: 1,
        mount_params: params,
-       connect_params: get_connect_params(socket) || %{},
-       connect_info: get_connect_info(socket) || %{},
        test_pid: session["test_pid"],
        connected?: connected?(socket),
        on_handle_params: on_handle_params && :erlang.binary_to_term(on_handle_params)